pyproject.toml
[tool.poetry]
name = "adoptive-market-crawler"
version = "0.1.0"
description = "市場分析のための情報収集とナレッジ化システム"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
neo4j = "^5.14.0"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
responses = "^0.24.1"
google-generativeai = ">=0.3.0,<0.4.0"
neo4j = "^5.27.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"

